refactor: draw graph에서 유틸 함수 분리
